refactor: adjust options with http.parser
